Add Number Features, first fixes for unit tests - still failing
