+++ utils/test/clearing_operations_tests/ClearingOperationsCreateGRGFromPointTestCase.py
@@ -30,24 +30,25 @@ import Configuration
 import UnitTestUtilities
 import DataDownload
 
-class ClearingOperationsCanvasAreaGRGTestCase(unittest.TestCase):
+class ClearingOperationsCreateGRGFromPointTestCase(unittest.TestCase):
 
     toolboxUnderTest = None # Set to Pro or ArcMap toolbox at runtime
 
+    pointTarget = None
     inputArea = None
     output = None
 
 
     scratchGDB = None
     def setUp(self):
-        if Configuration.DEBUG == True: print("         ClearingOperationsCanvasAreaGRGTestCase.setUp")
+        if Configuration.DEBUG == True: print("         ClearingOperationsCreateGRGFromPointTestCase.setUp")
 
         ''' Initialization needed if running Test Case standalone '''
         Configuration.GetLogger()
         Configuration.GetPlatform()
         ''' End standalone initialization '''
-        self.toolboxUnderTest = Configuration.clearingOperationsToolboxPath + \
-            Configuration.GetToolboxSuffix()
+
+        self.toolboxUnderTest = Configuration.clearingOperationsToolboxPath
 
         UnitTestUtilities.checkArcPy()
         DataDownload.runDataDownload(Configuration.clearingOperationsPath, \
@@ -57,41 +58,42 @@ class ClearingOperationsCanvasAreaGRGTestCase(unittest.TestCase):
             self.scratchGDB = UnitTestUtilities.createScratch(Configuration.clearingOperationsPath)
 
         # set up inputs
+        self.pointTarget = os.path.join(Configuration.clearingOperationsInputGDB, r"CenterPoint")
         self.inputArea = os.path.join(Configuration.clearingOperationsInputGDB, r"AO")
-
 
         UnitTestUtilities.checkFilePaths([Configuration.clearingOperationsPath])
 
-        UnitTestUtilities.checkGeoObjects([Configuration.clearingOperationsInputGDB, self.toolboxUnderTest, self.scratchGDB, self.inputArea])
+        UnitTestUtilities.checkGeoObjects([Configuration.clearingOperationsInputGDB, self.toolboxUnderTest, self.scratchGDB, self.pointTarget, self.inputArea])
 
     def tearDown(self):
-        if Configuration.DEBUG == True: print("         ClearingOperationsTestCase.tearDown")
+        if Configuration.DEBUG == True: print("         ClearingOperationsCreateGRGFromPointTestCase.tearDown")
         UnitTestUtilities.deleteScratch(self.scratchGDB)
 
-    def testClearingOperationsAreaGRG(self):
-        if Configuration.DEBUG == True:print(".....ClearingOperationsCanvasAreaGRGTestCase.testClearingOperationsAreaGRG")
+    def testClearingOperationsPointTarget(self):
+        if Configuration.DEBUG == True:print(".....ClearingOperationsCreateGRGFromPointTestCase.testClearingOperationsPointTarget")
         print("Importing toolbox...")
         arcpy.ImportToolbox(self.toolboxUnderTest)
         arcpy.env.overwriteOutput = True
 
-		
         #inputs
+        numCellsH = 9
+        numCellsV = 9
         cellWidth = 100
         cellHeight = 100
-        cellunits = "Meters"
+        cellUnits = "Meters"
         labelStart = "Lower-Left"
         labelStyle = "Alpha-Numeric"
-        output = os.path.join(self.scratchGDB, "grg")
+        output = os.path.join(self.scratchGDB, "ptTarget")
 
         #Testing
-        runToolMsg="Running tool (Canvas Area GRG)"
+        runToolMsg="Running tool (Point Target)"
         arcpy.AddMessage(runToolMsg)
         Configuration.Logger.info(runToolMsg)
 
         try:
-			# Calling the CanvasAreaGRG_ClearingOperations Script Tool
-            arcpy.CanvasAreaGRG_ClearingOperations(self.inputArea, cellWidth, cellHeight, cellunits, labelStart, labelStyle, output)
-			
+            # Calling the PointTargetGRG_ClearingOperations Script Tool
+            arcpy.CreateGRGfromPoint_clrops(self.pointTarget, numCellsH, numCellsV, cellWidth, cellHeight, \
+                "Meters", None, labelStart, labelStyle, output)
         except arcpy.ExecuteError:
             UnitTestUtilities.handleArcPyError()
         except:
@@ -100,7 +102,20 @@ class ClearingOperationsCanvasAreaGRGTestCase(unittest.TestCase):
         result = arcpy.GetCount_management(output)
         count = int(result.getOutput(0))
         print("number features: " + str(count))
-        self.assertEqual(count, 40)
+        self.assertGreaterEqual(count, 80)
+
+        #TODO: determine if should be added back in
+        #cursor = arcpy.da.SearchCursor(output, 'SHAPE@')
+        #cursor2 = arcpy.da.SearchCursor(self.pointTarget, 'SHAPE@')
+        #ptRow = cursor2.next()
+        #ptGeo = ptRow[0]
+        #intersect = False
+        #for row in cursor:
+        #    intersectFeat= row[0].touches(ptGeo)
+        #    #print(str(intersectFeat))
+        #    if intersectFeat == True:
+        #        intersect = True
+        #self.assertTrue(intersect)
 
 if __name__ == "__main__":
     unittest.main()
